Consolidate datastore with fallbacks logic
