+++ lib/msf/core/modules/external/templates/single_scanner.erb
@@ -23,7 +23,7 @@ class MetasploitModule < Msf::Auxiliary
   def run_host(ip)
     print_status("Running for #{ip}...")
     rhost = datastore.delete('RHOST')
-    datastore.remove_option('RHOST') if self.datastore.is_a?(Msf::DataStoreWithFallbacks)
+    datastore.remove_option('RHOST') if self.datastore.is_a?(Msf::DataStore)
     datastore['rhost'] = rhost
     execute_module(<%= meta[:path] %>)
   end
