+++ lib/msf/core/rhosts_walker.rb
@@ -65,7 +65,7 @@ module Msf
       return unless block_given?
 
       parse(@value, @datastore).each do |result|
-        block.call(result) if result.is_a?(Msf::DataStore) || result.is_a?(Msf::DataStoreWithFallbacks)
+        block.call(result) if result.is_a?(Msf::DataStore)
       end
 
       nil
@@ -99,7 +99,7 @@ module Msf
     # @return [Boolean] True if all items are valid, and there are at least some items present to iterate over. False otherwise.
     def valid?
       parsed_values = parse(@value, @datastore)
-      parsed_values.all? { |result| result.is_a?(Msf::DataStore) || result.is_a?(Msf::DataStoreWithFallbacks) } && parsed_values.count > 0
+      parsed_values.all? { |result| result.is_a?(Msf::DataStore) } && parsed_values.count > 0
     rescue StandardError => e
       elog('rhosts walker invalid', error: e)
       false
