+++ lib/msf/core/rpc/v10/rpc_core.rb
@@ -64,7 +64,7 @@ class RPC_Core < RPC_Base
   # @example Here's how you would use this from the client:
   #  rpc.call('core.unsetg', 'MyGlobal')
   def rpc_unsetg(var)
-    if framework.datastore.is_a?(Msf::DataStoreWithFallbacks)
+    if framework.datastore.is_a?(Msf::DataStore)
       framework.datastore.unset(var)
     else
       framework.datastore.delete(var)
