+++ lib/msf/ui/console/module_option_tab_completion.rb
@@ -17,10 +17,10 @@ module Msf
         #   stage since the command itself has been completed.
         def tab_complete_datastore_names(datastore, _str, _words)
           keys = (
-            Msf::DataStoreWithFallbacks::GLOBAL_KEYS +
+            Msf::DataStore::GLOBAL_KEYS +
               datastore.keys
           )
-          keys.concat(datastore.options.values.flat_map(&:fallbacks)) if datastore.is_a?(Msf::DataStoreWithFallbacks)
+          keys.concat(datastore.options.values.flat_map(&:fallbacks)) if datastore.is_a?(Msf::DataStore)
           keys.uniq! { |key| key.downcase }
           keys
         end
