+++ spec/support/shared/examples/msf/module/data_store.rb
